refartor(#24): 전체적인 nx 구조 개선
