+++ apps/web/src/app/layout.tsx
@@ -1,4 +1,9 @@
 import './global.css';
+import {
+  LayoutPage,
+  ServerCreateProvider,
+  ServerJoinedTeamProvider,
+} from '@meezy/ui';
 
 export const metadata = {
   title: 'Welcome to web',
@@ -12,7 +17,11 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
-      <body>{children}</body>
+      <body>
+        <ServerCreateProvider>
+          <ServerJoinedTeamProvider>{children}</ServerJoinedTeamProvider>
+        </ServerCreateProvider>
+      </body>
     </html>
   );
 }
