+++ apps/web/src/app/main/joined/TeamJoined.tsx
@@ -12,6 +12,7 @@ import {
   ChatRoom,
 } from '@meezy/ui';
 import { colors, typography } from '@meezy/ui';
+import { roomsrcList, userList, sidebarList } from '../context/list';
 
 export function TeamJoined() {
   const [isModalOpen, setIsModalOpen] = useState(false);
@@ -25,58 +26,6 @@ export function TeamJoined() {
   const onCloseModal = () => {
     setIsModalOpen(false);
   };
-
-  const sidebarList: {
-    team_id: number;
-    room_name: string;
-    type: 'ROOM' | 'MEMBER' | null;
-    create_at: null;
-  }[] = [
-    {
-      team_id: 1,
-      room_name: '대화',
-      type: 'ROOM',
-      create_at: null,
-    },
-    {
-      team_id: 2,
-      room_name: '멤버',
-      type: 'MEMBER',
-      create_at: null,
-    },
-  ];
-
-  const roomsrcList = [
-    {
-      room_id: 1,
-      team_id: 1,
-      room_name: '환영',
-      create_at: null,
-    },
-    {
-      room_id: 2,
-      team_id: 1,
-      room_name: '잡담',
-      create_at: null,
-    },
-  ];
-
-  const userList = [
-    {
-      user_id: 1,
-      team_id: 2,
-      user_name: '정명우',
-      create_at: null,
-      img: null,
-    },
-    {
-      user_id: 2,
-      team_id: 2,
-      user_name: '김효현',
-      create_at: null,
-      img: null,
-    },
-  ];
 
   const roomNameFind =
     roomsrcList.find((room) => room.room_id === selectedRoomId)?.room_name ??
