+++ apps/web/src/app/main/page.tsx
@@ -1,0 +1,5 @@
+import { LayoutPage } from '@meezy/ui';
+
+export default function MainPage() {
+  return <LayoutPage />;
+}
