+++ apps/web/src/assets/noTeam_Receive.png
