+++ libs/ui/src/index.ts
@@ -2,14 +2,14 @@ export * from './design';
 export * from './lib/input';
 export * from './lib/CalendarMockup';
 export * from './lib/header';
-export * from './lib/sidebar';
-export * from './lib/serverModal';
-export * from './lib/LayoutPage';
+export * from './lib/sidebar/sidebar';
+export * from './lib/models/serverModal';
+export * from '../../../apps/web/src/app/main/LayoutPage';
 export * from './context/ServerCreateProvider';
 export * from './context/ServerJoinedTeamProvider';
-export * from './lib/TeamJoined';
-export * from './lib/joinedSidebar';
+export * from '../../../apps/web/src/app/main/joined/TeamJoined';
+export * from './lib/sidebar/joinedSidebar';
 export * from './lib/DashboardCard';
 export * from './lib/ParticipationChart';
-export * from './lib/joinedModel';
+export * from './lib/models/joinedModel';
 export * from './lib/chatRoom';
