--- libs/ui/src/lib/LayoutPage.tsx
@@ -1,10 +1,0 @@
-'use client';
-
-import { TeamJoin } from './TeamJoin';
-import { TeamJoined } from './TeamJoined';
-import { useServerJoinedTeam } from '../context/ServerJoinedTeamProvider';
-
-export function LayoutPage() {
-  const { joined } = useServerJoinedTeam();
-  return <>{joined ? <TeamJoined /> : <TeamJoin />}</>;
-}
