+++ libs/ui/src/lib/index.ts
@@ -1,10 +1,10 @@
 export * from './input';
 export * from './CalendarMockup';
 export * from './header';
-export * from './sidebar';
-export * from './serverModal';
-export * from './LayoutPage';
-export * from './joinedSidebar';
+export * from './sidebar/sidebar';
+export * from './models/serverModal';
+export * from '../../../../apps/web/src/app/main/LayoutPage';
+export * from './sidebar/joinedSidebar';
 export * from './DashboardCard';
 export * from './ParticipationChart';
-export * from './joinedModel'
+export * from './models/joinedModel';
