+++ libs/ui/src/lib/sidebar/index.ts
@@ -1,0 +1,2 @@
+export * from './sidebar';
+export * from './joinedSidebar';
